requirements.txt

# Core Dependencies
python-telegram-bot>=20.7
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.2

# AI and Language Models
openai>=1.3.0

# Memory Systems
chromadb>=0.4.22
supabase-py>=2.3.0

# Web Processing
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Utilities
uuid>=1.30
numpy>=1.24.3
tenacity>=8.2.3
